+++ src/index.html
@@ -7,7 +7,7 @@
 		<title>Tommy's website</title>
 	</head>
 	<body>
-		<noscript>You need to enable JavaScript to run this app.</noscript>
+		<noscript>Please enable JavaScript to run this app.</noscript>
 		<div id="root"></div>
 		<script type="module" src="index.tsx"></script>
 	</body>
