+++ src/pages/Portfolio.tsx
@@ -1,7 +1,9 @@
 import React from 'react'
-import PreviewTab from '../components/PreviewTab'
+import PreviewItem from '../components/PreviewItem'
 import { useLoaderData } from 'react-router-dom'
 import { Post } from '../data/model/Post'
+
+import '../style/portfolio.scss'
 
 const Portfolio = () => {
   const { portfolio } = useLoaderData() as { portfolio: Post[] }
@@ -10,17 +12,21 @@ const Portfolio = () => {
       <h1>Portfolio</h1>
       Below is a selection of a few projects I have worked on. Source code for all these and other projects can be found
       on on my GitHub. Enjoy!
-      <PreviewTab imgUri="raven-game-engine-preview" title="Raven Game Engine" route="/portfolio/raven-game-engine">
+      <PreviewItem
+        imgUri="/assets/RavenEngineEditor.png"
+        title="Raven Game Engine"
+        route="/portfolio/raven-game-engine"
+      >
         Leading a 5-person team to build a C++ game engine from scratch.
-      </PreviewTab>
-      <PreviewTab
-        imgUri="unity-protein-visualiser-preview"
+      </PreviewItem>
+      <PreviewItem
+        imgUri="/assets/Ubiquitin.png"
         title="Unity Protein Visualiser"
         route="/portfolio/unity-protein-visualiser"
         justifyImgRight
       >
         Exploring the application of game technologies in research by building a protein visualiser in Unity.
-      </PreviewTab>
+      </PreviewItem>
     </>
   )
 }
