+++ src/style/root.scss
@@ -46,6 +46,7 @@ a:hover, a:focus {
   padding: 1vh;
   align-items: center;
 }
+
 .nav-item:first-child {
   list-style-type: none;
   margin-right: auto;
