+++ vite.config.ts
@@ -7,6 +7,7 @@ import remarkMdxFrontmatter from 'remark-mdx-frontmatter'
 // https://vitejs.dev/config/
 export default defineConfig({
   root: './src',
+  publicDir: '../public',
   build: {
     outDir: '../dist',
     rollupOptions: {
